+++ apps/appcontainer-node/app/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.39.8-alpha.4](https://github.com/nrkno/sofie-package-manager/compare/v1.39.8-alpha.3...v1.39.8-alpha.4) (2023-01-17)
+
+**Note:** Version bump only for package @appcontainer-node/app
+
+
+
+
+
+## [1.39.8-alpha.3](https://github.com/nrkno/sofie-package-manager/compare/v1.39.8-alpha.2...v1.39.8-alpha.3) (2023-01-13)
+
+**Note:** Version bump only for package @appcontainer-node/app
+
+
+
+
+
+## [1.39.8-alpha.2](https://github.com/nrkno/sofie-package-manager/compare/v1.39.8-alpha.1...v1.39.8-alpha.2) (2023-01-12)
+
+**Note:** Version bump only for package @appcontainer-node/app
+
+
+
+
+
 ## [1.39.8-alpha.1](https://github.com/nrkno/sofie-package-manager/compare/v1.39.8-alpha.0...v1.39.8-alpha.1) (2023-01-12)
 
 **Note:** Version bump only for package @appcontainer-node/app